requirements.txt
requests>=2.28.0
openai>=1.0.0
pyyaml>=6.0
python-dateutil>=2.8.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
flask>=2.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pytest>=7.0.0
